MudThemingProvider:  MudThemeProvider without MudPopoverProvider
